feat(components): Opereation component
This component provides a simple way to show "operations" made by your
code sample.

- Component path: src/components/Operation.tsx
- Usage: you have to provide props like title, description and codeURL
  to use it.

- Props:

	* title: string
	* description: string,
	* codeURL: string

The codeURL is used to fetch the file that is going to be read and
insert it to the monaco editor.

For the description you have to send text but when you want a line break
you have to specify it with the # character.

Example:

	Jorge now wants to sleep.#
	Jorge shouldn't do that until tomorrow.#